proyecto-transporte/requirements.txt
# Pipeline de datos
pandas>=2.0
numpy>=1.24
pyyaml>=6.0

# Modelado
lightgbm>=4.0
scikit-learn>=1.3
optuna>=3.0
joblib>=1.3

# Registro de experimentos
mlflow>=2.0

# API REST
fastapi>=0.100
uvicorn>=0.23
pydantic>=2.0
requests>=2.31

# Dashboard
streamlit>=1.25
plotly>=5.0

# Notebooks
jupyter